Create ISY994i.psm1 file
Requires Powershell version 5

This module is a collection of powershell functions which leverage the ISY994i's published REST and SOAP interfaces.

After setting the execution policy in powershell (for example 'Set-ExecutionPolicy Unrestricted' in an elevated session) to something other than restricted create the following directory: 

$HOME\Documents\WindowsPowerShell\Modules\ISY994i

 and then place the ISY994i.psm1 file in it.

Then type 'Get-Help New-ISYSettings -full'. This will explain the 2 ways to establish the settings.
The easiest way to begin is to then type 'New-ISYSettings -usegui'.  This  will launch a form to set the the defaults.

After installation to get all the functions available type:

Get-Command -Module ISY994i  

From this list you can Get-Help for each command.

Note the admin console can do everything this module can and more. This just provides a way to use the command line for many actions.